+++ src/main/java/com/example/geektrust/strategy/AdultFareStrategy.java
@@ -3,12 +3,12 @@ package com.example.geektrust.strategy;
 public class AdultFareStrategy implements FareStrategy{
 
   @Override
-  public double getBaseFare() {
-    return 200.0;
+  public int getBaseFare() {
+    return 200;
   }
 
   @Override
-  public double getDiscountedFare() {
-    return getBaseFare() * 0.5;
+  public int getDiscountedFare() {
+    return (int) (getBaseFare() * 0.5);
   }
 }
