+++ src/main/java/com/example/geektrust/strategy/FareStrategy.java
@@ -1,6 +1,6 @@
 package com.example.geektrust.strategy;
 
 public interface FareStrategy {
-  double getBaseFare();
-  double getDiscountedFare();
+  int getBaseFare();
+  int getDiscountedFare();
 }
