+++ src/main/java/com/example/geektrust/strategy/SeniorCitizenFareStrategy.java
@@ -3,13 +3,13 @@ package com.example.geektrust.strategy;
 public class SeniorCitizenFareStrategy implements FareStrategy {
 
   @Override
-  public double getBaseFare() {
-    return 100.0;
+  public int getBaseFare() {
+    return 100;
   }
 
   @Override
-  public double getDiscountedFare() {
-    return getBaseFare() * 0.5;
+  public int getDiscountedFare() {
+    return (int) (getBaseFare() * 0.5);
   }
 
 }
